# Conflicts:
#	.eslintignore
#	components/alert/alert.examples.html
#	components/autofocus-ng/autofocus-ng.examples.html
#	components/button-group/button-group.examples.html
#	components/button-set/button-set.examples.html
#	components/button-toolbar/button-toolbar.js
#	components/button/button.examples.html
#	components/caret/caret.js
#	components/checkbox/checkbox.examples.html
#	components/checkbox/checkbox.js
#	components/contenteditable/contenteditable.js
#	components/date-picker/date-picker.js
#	components/dialog-ng/dialog-ng.examples.html
#	components/dialog/dialog.examples.html
#	components/dialog/dialog.js
#	components/docked-panel-ng/docked-panel-ng.js
#	components/dropdown/dropdown.examples.html
#	components/error-bubble/error-bubble.js
#	components/footer/footer.js
#	components/form-ng/form-ng.examples.html
#	components/form/form.examples.html
#	components/header/header.examples.html
#	components/input/input.examples.html
#	components/link/link.examples.html
#	components/list/list.js
#	components/list/list__item.js
#	components/list/list__separator.js
#	components/list/list__title.js
#	components/panel/panel.js
#	components/popup-menu/popup-menu.js
#	components/promised-click-ng/promised-click-ng.js
#	components/react-ng/react-ng.js
#	components/select/select.examples.html
#	components/select/select.js
#	components/tooltip/tooltip.js
#	docs/breaking-changes.md
#	gemini/screens/Icon/All icons/all icons/chrome.png
#	gemini/screens/Icon/All icons/all icons/firefox.png
#	gemini/screens/Icon/All icons/all icons/ie.png
#	gemini/screens/Select/Input based select/selectPopup/chrome.png
#	gemini/screens/Select/Input based select/selectPopup/firefox.png
#	gemini/screens/Select/Input based select/selectPopup/ie.png
#	gemini/screens/Select/Select with filter/selectPopup/chrome.png
#	gemini/screens/Select/Select with filter/selectPopup/firefox.png
#	gemini/screens/Select/Select with filter/selectPopup/ie.png
#	package.json
#	packages/ring-ui-docs/components/header.js
#	packages/ring-ui-docs/components/index.css
#	packages/ring-ui-docs/components/nav.js
#	packages/ring-ui-docs/package.json
#	packages/ring-ui-generator/app/templates/src/components/app/app-root.js

+++ generator_test.sh
@@ -23,6 +23,10 @@ function test {
   cleanup
 }
 
+if [ -v IS_DEFAULT_BRANCH ]
+then
+  exit 0
+fi
 
 rm -rf test_gen
 test app
